More aggressive job deletes in KubernetesJobOperator (#156)
We have received errors while trying to delete successful jobs in Airflow. That takes a successful job and turns it into a failure, breaking the workflow. Note that in the logs below there is a 5 mintue delay between attempts, due to the timeout of `delete`, and that is still not enough.

```
error: timed out waiting for "next-best-2cdd74be16e44562" to be synced
[2019-04-15 04:01:11,239] {{kubernetes_operator.py:176}} INFO - Retrying check_output because Command '['kubectl', 'delete', 'job', u'next-best-2cdd74be16e44562']' returned non-zero exit status 1
error: timed out waiting for "next-best-2cdd74be16e44562" to be synced
[2019-04-15 04:06:12,415] {{kubernetes_operator.py:176}} INFO - Retrying check_output because Command '['kubectl', 'delete', 'job', u'next-best-2cdd74be16e44562']' returned non-zero exit status 1
error: timed out waiting for "next-best-2cdd74be16e44562" to be synced
[2019-04-15 04:11:13,593] {{kubernetes_operator.py:176}} INFO - Retrying check_output because Command '['kubectl', 'delete', 'job', u'next-best-2cdd74be16e44562']' returned non-zero exit status 1
error: timed out waiting for "next-best-2cdd74be16e44562" to be synced
```

This change instead forces deletes after 2 minutes, and also ignores double-delete because don't care. Also, we don't wait for finalizers.