Updating requirement requests>=2.22.0 (#203) (#204)
* Updating requirement requests>=2.22.0

We had been pulling in only >=2.5.1. Previous to October 3rd, we had been pulling in some version >=2.18.1, though the exact version is unknown. Since October 3rd, we have been pulling in 2.12.4. All versions older 2.18.1 do not present the `Response` object as a context manager. We explicitly use that object as a context manager, so `AppEngineOperatorSync` has been failing in all DAGs since then. This has lead to the list of active namespaces falling out of date.

This change explicitly requests a recent version of `requests`, which has been tested in QA and works properly. Additionally, the list of of active namespaces has been manually updated in `airflow-prod` to tide us over until this change is deployed.

https://bluecore.atlassian.net/browse/PROD-20565