Adding hint to status checking queries
In a completely uncharacteristic and disappoing move, MySQL's optimizer fails to make good decisions about index choices. When there are no task IDs provided to the query, it will use `ti_dag_state`, which is highly selective when the state is `running`. We have added a hint. While forcing the index is not a generally safe assumption, the only place `airflow.models.DAG.get_num_task_instances` is called is in airflow.jobs.SchedulerJob._find_executable_task_instances(), which is looking specifically for running states, and `airflow.models.DAG.concurrency_reached` also looks specifically for `running`. The hint should keep the number of inspected rows down to a dull roar.

[sqlalchemy.orm.query.Query.with_hint docs](https://docs.sqlalchemy.org/en/13/orm/query.html#sqlalchemy.orm.query.Query.with_hint)